dDowa Household Latrine survey 2020

This dataset documents household-level sanitation information collected in the Dowa District. Each row represents an individual household's sanitation setup, usage patterns, perceptions, and maintenance behaviors. The goal is to understand the quality, sustainability, and user satisfaction with existing latrines in rural and peri-urban communities.

Usage

dowalatrinesurvey

Format

A tibble with 89 rows and 80 variables

submitted_on

community_type

latitude

longitude

traditional_authority

district

familiar_sanitation_types

latrines_in_community

sanitation_system_home

photo_sanitation_facility

sanitation_location_to_home

purchase_installation_process

latrine_installed

system_duration_magnitude

system_duration_unit

latrine_builder

facility_in_use

is_pit_lined

pit_lining_material

other_waste_disposed

other_waste_specified

biogas_generation

toilet_build_cost

cost_affordability

reasonable_cost_magnitude

reasonable_cost_unit

install_duration_magnitude

install_duration_unit

latrine_usage_process

safe_for_vulnerable

adapted_for_disabled

men_use_for_urination

liked_features

liked_features_other

sit_or_squat

currently_do

used_at_night

has_light

bring_own_light

bring_own_light_comments

odor_control

has_vent_pipe

prefer_less_odor

satisfaction_with_latrine

issues_with_animals

animal_issues_other

ideal_latrine_description

latrine_within_30m_water

latrine_distance_to_water

water_source

flush_with_water

latrine_description

soil_description

soil_description_other

latrine_user_count

latrine_built_year

expected_lifespan_magnitude

expected_lifespan_unit

privacy_enclosure_other_uses

above_ground_maintenance

above_ground_maintenance_other

above_ground_cleaning

uses_cleaning_chemicals

maintenance_frequency

maintenance_by

maintenance_by_other

maintenance_cost

cleaning_major_issue

pit_latrine_condition

is_filled_up

is_replacement

previous_filled_count

fill_up_speed

is_abandoned

reason_for_abandonment

is_emptied

why_not_emptied

why_not_emptied_other

emptying_frequency

emptying_method